requirements.txt
mysql-connector-python
fpdf
matplotlib

